requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
sqlalchemy==2.0.17
pydantic==2.1.0
requests==2.31.0
python-dotenv==1.0.0
asyncpg==0.26.0
